POC: Execution via an `xarray.Dataset` Table. (#1)
This is a proof of concept for Qarray. Here, I make use of sqlglot's
built-in SQL executor by making a custom Xarray table. The main function
of this table is to translate a multi-dimensional array into a flat
table index.

This is only tested with a simple select-and-limit. More tests are
needed.