Dataframe converter (#9)
This change provides a few humble functions to try to adapt the Xarray
model to Dask's dataframe model. The conversion is more or less an
itertools.product and index operation. The translation to dataframes
honor's Xarray's chunks.

There area a few next steps. From here, we can add dask-sql to this
module and see how it works on real SQL queries.