Added support for Dask Dataframe chunks. (#24)
Users can specify chunks to break up the Xarray Dataset that may differ
from the Dataset's own chunking mechanism. This is useful, for example,
when working with Zarr. Here, you can set the Zarr chunks to be big and
the unravel chunks to be small (more partitions in the resulting
DataFrame). This can help limit the overall memory copying during
unraveling.