NumPy Cross-Product for Coords. (#25)
This patch provides a performance improvement to the current slowest
part of `unravel_unbounded()`. Instead of using itertools and looping to
create the cross product, we do it efficiently in NumPy via
`np.meshgrid()`. This also includes new performance profiles.

I discovered here a better idea of what chunk sizes should be! Now, we
can transform bigger Datasets to DataFrames with the right chunks.