Rename to xarray-sql. (#41)
This is Qarray no longer! Now, it's Xarray-sql (correct capitalization
TBD). The name change is due to the fact that `Qarray` is too similar to
other packages in PyPi (i.e. `Xarray`) and is not allowed to be
published. https://pypi.org/help/#project-name This is OK in my book;
"Qarray" was not a perfect name, either.