More efficient `read_xarray` implementation with `from_map_batched` (#63)
I created a lightweight iterator over batches called `from_map_batched`,
which uses a pyarrow RecordBatchReader to stream in batches one by one.
This is not the ideal of a streaming dataset, which after lots of
research, requires a TabeProvider via a FFI. But, it is a
memory-efficient improvement over the present implementation.

This PR provides other small refactorings and reformattings.